+++ srcs/invite.cpp
@@ -6,14 +6,14 @@
 /*   By: caguillo <caguillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 15:13:27 by caguillo          #+#    #+#             */
-/*   Updated: 2025/04/24 15:06:03 by caguillo         ###   ########.fr       */
+/*   Updated: 2025/04/24 23:22:29 by caguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Server.hpp"
 
 // :inviter INVITE <invitee> <channel> // INVITE no parameter --> not standard
-void Server::invite(std::vector<std::string>& tab_msg, int clt_idx, int tab_idx)
+void Server::invite(std::vector<std::string>& tab_msg, int clt_idx, size_t tab_idx)
 {
     size_t i = tab_idx + 1; // invitee
