+++ srcs/part.cpp
@@ -6,14 +6,14 @@
 /*   By: caguillo <caguillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/06 18:17:45 by caguillo          #+#    #+#             */
-/*   Updated: 2025/04/24 15:09:17 by caguillo         ###   ########.fr       */
+/*   Updated: 2025/04/24 23:23:11 by caguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Server.hpp"
 
 // PART <channel>{,<channel>} [ :<message> ]
-void Server::part(std::vector<std::string>& tab_msg, int clt_idx, int tab_idx)
+void Server::part(std::vector<std::string>& tab_msg, int clt_idx, size_t tab_idx)
 {
     size_t i = tab_idx + 1;
     std::string msg_replied;
